+++ extension/src/popup/components/Menu/styles.scss
@@ -1,0 +1,31 @@
+.Settings {
+  &--slide-out {
+    background: var(--pal-background-primary);
+    height: 100%;
+    left: 0;
+    margin-left: -100%;
+    max-width: 456px;
+    overflow: hidden;
+    padding: 2rem 1.5rem;
+    position: absolute;
+    top: 0;
+    transition: margin var(--anim-transition-default);
+    width: 100%;
+    z-index: var(--z-index--settings);
+
+    &--open {
+      margin-left: 0;
+    }
+  }
+
+  &--nav-link {
+    display: flex;
+    justify-content: space-between;
+    line-height: 1.5rem;
+    padding: 0.75rem 0;
+
+    a {
+      color: var(--pal-text-primary);
+    }
+  }
+}
