+++ extension/src/popup/styles.scss
@@ -1,0 +1,3 @@
+:root {
+  --z-index--settings: 10;
+}
